Slice lexer.rs into parts
It certainly holds too much in there. Diagnostic processing will certainly
improve and be extended in the future. And the users would certainly like
to import token types without importing the lexer which produces them.
